+++ app/commands/homebrew_context/daggerheart/change_speciality_command.rb
@@ -3,10 +3,6 @@
 module HomebrewContext
   module Daggerheart
     class ChangeSpecialityCommand < BaseCommand
-      include Deps[
-        refresh_user_data: 'services.homebrews_context.refresh_user_data'
-      ]
-
       use_contract do
         config.messages.namespace = :homebrew_speciality
 
@@ -20,8 +16,6 @@ module HomebrewContext
 
       def do_persist(input)
         input[:speciality].update!(input.slice(:name))
-
-        refresh_user_data.call(user: input[:speciality].user) if input[:speciality].user
 
         { result: input[:speciality] }
       end